Update Hjson to new "fix" object field and re-arrange
Converted all existing fix fields to the new object pattern.

Added manual instructions for configs where this was previously marked
as “TODO”. This satisfies the Hjson portion of issue #64.

Re-arranged most or all of the config checks so that all of the
check-related fields are in order, and generally put things in a more
logical order. When I implement #41, this will now be substantially
easier.