osx-check.json typos in prompts: disaled >disabled
Despite the warning regarding autogeneration of the osx-check.json
from its .hjson source; there are four instances of this typo in
the json which are not present in the *.hjson file.  Also the
hjson_to_json.py utility has an unsatisfied dependency on some
hjson module.

TODO: include references to the hjson module