Break tests into list with individual pass and fail states
This commit changes the Hjson syntax to support a list of check tests,
each of which can pass or fail early and prevent subsequent tests from
being evaluated. This simplifies the syntax for commands that have
simple prerequisites such as “is this app even installed?”

An early fail state also helps avoid running checks with sudo privy
unnecessarily.

QA: Did lots of manual testing, running the app with debug print mode
enabled and carefully reviewing output. Also confirmed that tests work
correct when Google Chrome is installed or not installed, now that
Chrome-related checks are now making use of this change.