fix the disabling of AirDrop
This applies two fixes to the AirDrop setting:

1) Based on https://github.com/kristovatlas/osx-config-check/pull/89 by
@hypn this writes defaults for only the current user
2) Through testing on a live system I determined that even a system
with AirDrop enabled will not necessarily have the
`com.apple.NetworkBrowser` domain defined, but if we do define it
manually and set DisableAirDrop to true, then AirDrop is disabled. Note
that AirDrop generally requires iCloud sign-in before it can be enabled.

The new defaults_write_ignore_missing API function could be used for
other writes to defaults if needed.