Log debug data by default
Added warning to message that explains where log was written, now that
it has more sensitive info by default.

QA: Ran the app in variety of modes and confirmed that the everything
works as expected. (TODO: This would be a good thing to check in the
future by supplying a “debug” config list and confirming the stdout and
log file in a variety of modes.)